cr-api-server/pyproject.toml
[tool.poetry]
name = "cr-api-server"
version = "0.1.0"
description = ""
authors = ["pelle@live.cn"]

[tool.poetry.dependencies]
python = "^3.8"
redis = "^3.5.3"
Flask = "^2.0.1"
gunicorn = "^20.1.0"
marshmallow = "^3.12.1"
webargs = "^8.0.0"
PyJWT = "^2.3.0"
Flask-Mail = "^0.9.1"
Flask-SQLAlchemy = "^2.5.1"
PyMySQL = "^1.0.2"


[tool.poetry.dev-dependencies]
autopep8 = "^1.6.0"
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
